Relocate ChallengeManager to services and rename to ChallengePrompter
- Moved `ChallengeManager.kt` from `ui.dashboard` to `services` package.
- Renamed `ChallengeManager` to `ChallengePrompter`.
- Updated package declaration and imports in `ChallengePrompter.kt`.
- Updated imports and variable names in `DashboardActivity.kt` and `BellDashboardFragment.kt` to reference the new class.
